Lockscreen charging info: show decimal in battery temperature

+++ packages/SettingsLib/src/com/android/settingslib/fuelgauge/BatteryStatus.java
@@ -55,13 +55,13 @@ public class BatteryStatus {
     public final int maxChargingCurrent;
     public final int maxChargingVoltage;
     public final int maxChargingWattage;
-    public final int temperature;
+    public final float temperature;
     public final boolean present;
     public final boolean oemFastChargeStatus;
 
     public BatteryStatus(int status, int level, int plugged, int health,
             int maxChargingCurrent, int maxChargingVoltage,
-            int maxChargingWattage, int temperature,
+            int maxChargingWattage, float temperature,
             boolean present, boolean oemFastChargeStatus) {
 
         this.status = status;
